+++ setup/site/config/application.rb
@@ -30,10 +30,7 @@ Application = Utopia::Application.build do
 	}
 	
 	use Utopia::Redirection::DirectoryIndex
-	
-	use Utopia::Redirection::Errors, {
-		404 => "/errors/file-not-found"
-	}
+	use Utopia::Redirection::Errors, 404 => "/errors/file-not-found"
 	
 	use Utopia::Session,
 		expires_after: 3600 * 24,