+++ test/utopia/redirection.rb
@@ -23,7 +23,7 @@ describe Utopia::Redirection do
 	
 	let(:middleware) do
 		Utopia::Application.build(Protocol::HTTP::Middleware.for{|request|
-			case request.path_info
+			case request.url.path.encoded
 			when "/error"
 				Utopia::Response.text("File not found :(", 200)
 			when "/teapot"
@@ -34,13 +34,10 @@ describe Utopia::Redirection do
 		}) do
 			use Utopia::Redirection::Rewrite, {"/" => "/welcome/index"}
 			use Utopia::Redirection::DirectoryIndex
-			use Utopia::Redirection::Errors, {
-				404 => "/error",
-				418 => "/teapot"
-			}
 			use Utopia::Redirection::Moved, "/a", "/b"
 			use Utopia::Redirection::Moved, "/hierarchy/", "/hierarchy", flatten: true
 			use Utopia::Redirection::Moved, "/weird", "/status", status: 333
+			use Utopia::Redirection::Errors, 404 => "/error", 418 => "/teapot"
 		end
 	end
 	
@@ -84,10 +81,29 @@ describe Utopia::Redirection do
 		expect(last_response.read).to be == "File not found :("
 	end
 	
+	it "bypasses request redirections for internal error documents" do
+		application = Utopia::Application.build(Protocol::HTTP::Middleware.for do |request|
+			if request.url.path.encoded == "/error"
+				Utopia::Response.text("Internal error document")
+			else
+				Utopia::Response[404, {}, []]
+			end
+		end) do
+			use Utopia::Redirection::Rewrite, {"/error" => "/redirected"}
+			use Utopia::Redirection::Errors, 404 => "/error"
+		end
+		
+		response = application.call(Protocol::HTTP::Request["GET", "/missing"])
+		
+		expect(response.status).to be == 404
+		expect(response.headers["location"]).to be == nil
+		expect(response.read).to be == "Internal error document"
+	end
+	
 	it "closes the response replaced by an error document" do
 		events = []
 		application = Utopia::Application.build(Protocol::HTTP::Middleware.for do |request|
-			if request.path_info == "/error"
+			if request.url.path.encoded == "/error"
 				Utopia::Response[200, {}, tracked_body(:error, events)]
 			else
 				Utopia::Response[404, {}, tracked_body(:original, events)]
@@ -110,7 +126,7 @@ describe Utopia::Redirection do
 	it "closes both responses when the error document fails" do
 		events = []
 		application = Utopia::Application.build(Protocol::HTTP::Middleware.for do |request|
-			if request.path_info == "/error"
+			if request.url.path.encoded == "/error"
 				Utopia::Response[500, {}, tracked_body(:error, events)]
 			else
 				Utopia::Response[404, {}, tracked_body(:original, events)]
@@ -141,5 +157,15 @@ describe Utopia::Redirection do
 		
 		expect(last_response.status).to be == 333
 		expect(last_response.headers["location"]).to be == "/status"
+	end
+	
+	it "uses exact-path lookup" do
+		application = Utopia::Application.build do
+			use Utopia::Redirection::Rewrite, {"/files/" => "/exact"}
+		end
+		
+		response = application.call(Protocol::HTTP::Request["GET", "/files/"])
+		
+		expect(response.headers["location"]).to be == "/exact"
 	end
 end